Sketch of R-Tree insertion.
Looks like I'm spending more time in Docco and `README.md` so I'm not
adding as much to the commit log. You'll see writing if you view the
diff.

Added insert according to the R-Tree Guttman '84 paper. Have not gotten
it to work. Need to rework the Area object to perform this insertion.
Going to rename it `box` because `area` is a property name.

Have the concept of a Shape, but as I write this I realize it can be a
yet another property so that is outgoing.

See #19.