Upgrade Destructible and Interrupt.
Destructible was upgraded for Interrupt.

Interrupt was detrastically rewritten. Might want to include the stack
poking facilities from Interrupt into this module so that a stack trace
pokes into where a stream iterartor was constructed so I could
distinquish between an r-tree WAL faliure and a b-tree WAL failure, but
that is probably better accomplished through catch and wrap.

Found that calling `destroy` with an `Error` on a stream being used as
an async iterator will raise an exception instead of emitting an error
through the `EventEmitter` interface. Verified that this is the expected
behavior on an `fs.createReadStream()` stream.

https://github.com/bigeasy/interrupt/releases/tag/v11.0.0-alpha.3
https://github.com/bigeasy/destructible/releases/tag/v7.0.0-alpha.15

Closes #25.